Move internal/mod to daemon/internal/builder-next/worker/mod
