+++ daemon/internal/builder-next/worker/worker.go
@@ -16,9 +16,9 @@ import (
 	"github.com/containerd/platforms"
 	imageadapter "github.com/docker/docker/daemon/internal/builder-next/adapters/containerimage"
 	mobyexporter "github.com/docker/docker/daemon/internal/builder-next/exporter"
+	"github.com/docker/docker/daemon/internal/builder-next/worker/mod"
 	distmetadata "github.com/docker/docker/distribution/metadata"
 	"github.com/docker/docker/distribution/xfer"
-	"github.com/docker/docker/internal/mod"
 	"github.com/docker/docker/layer"
 	pkgprogress "github.com/docker/docker/pkg/progress"
 	"github.com/moby/buildkit/cache"
